requirements.txt
ryanair-py
pandas
streamlit
openpyxl